% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analysis}
\alias{analysis}
\title{Analyse Status and Trends of indicators}
\usage{
analysis(data = pillar_ecol_df, type = NULL)
}
\arguments{
\item{data}{pillar_ecol_df from the targets file}

\item{type}{a character of either 'trend' or 'status' indicating if the}
}
\value{
a vector of trend and status claims
}
\description{
This function automatically calculates status and trends
for a list of indicators used in the Marine Conservation
Target (MCT) app. This function looks for a shape file
in MarConsNetAnalaysis/data that is made with the function
make.names() of the area of interest.
}
\examples{
\dontrun{
trends <- analysis(data=pillar_ecol_df, type='trend')
}
}
