% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_method_comparison.R
\name{bio_method_comparison}
\alias{bio_method_comparison}
\title{Creates a method comparison plot}
\usage{
bio_method_comparison(datas = NULL, type = "bar")
}
\arguments{
\item{datas}{a data frame or list of data frames obtained from \code{get_project_data()}}

\item{type}{an argument indicating what the output should be. Options include
\code{bar}, which produces a bar chart, \code{venn}, which produces a Venn Diagram, or
\code{table}, which produces a table.}
}
\description{
This function creates a method comparison plot for
data collected in MPAs
}
\examples{
\dontrun{
library(MarConsNetData)
library(MarConsNetAnalysis)
d <- get_project_data(ids=c(1093, 642))
bio_method_comparison(datas=d)
}
}
