% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_species_accumulation.R
\name{bio_species_accumulation}
\alias{bio_species_accumulation}
\title{Creates a species accumulation curve}
\usage{
bio_species_accumulation(datas = NULL)
}
\arguments{
\item{datas}{a data frame or list of data frames obtained from \code{get_project_data()}}
}
\description{
This function creates a species accumulation plot for
data collected in MPAs
}
\examples{
\dontrun{
library(MarConsNetData)
library(MarConsNetAnalysis)
d <- get_project_data(ids=1093)
bio_species_accumulation(datas=d)
}
}
