% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_group_score.R
\name{calc_group_score}
\alias{calc_group_score}
\title{Calculate group scores}
\usage{
calc_group_score(
  df,
  grouping_var = NULL,
  score_var = "score",
  weight_var = "weight"
)
}
\arguments{
\item{grouping_var}{character string for the name of the grouping column in \code{df}. Use \code{NULL} (default) to calculate the overall score.}

\item{score_var}{character string for the name of the score column in \code{df}}

\item{weight_var}{character string for the name of the weight column in \code{df}}
}
\value{
data.frame with group scores
}
\description{
Calculate group scores
}
\examples{
df <- data.frame(objective=rep(c("Biodiversity",
                                 "Habitat",
                                 "Productivity"),
                               times=c(3,5,3)),
                 bins=c("Genetic Diversity",
                        "Species Diversity",
                        "Functional Diversity",

                        "Environmental Representativity",
                        "Key Fish Habitat",
                        "Connectivity",
                        "Uniqueness",
                        "Threats to Habitat",

                        "Biomass Metrics",
                        "Structure and Function",
                        "Threats to Productivity"),
                 score=runif(11,45,100),
                 weight=1,
                 area_name = "Random Example MPA")
# overall score
overall <- calc_group_score(df)

# group score
group <- calc_group_score(df, grouping_var = "objective")
}
