% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_in_sea_distance.R
\name{calc_in_sea_distance}
\alias{calc_in_sea_distance}
\title{in_sea_distance}
\usage{
calc_in_sea_distance(
  cellsize = 1000,
  bioregion,
  areas,
  units = "km",
  crs = paste0("+proj=aeqd +lon_0=", st_coordinates(st_centroid(bioregion))[1, 1],
    " +lat_0=", st_coordinates(st_centroid(bioregion))[1, 2],
    " +datum=WGS84 +units=m +no_defs")
)
}
\arguments{
\item{cellsize}{numeric with target cellsize in meters of hexagonal cells the distance between opposite edges. The edge length is cellsize/sqrt(3))}

\item{bioregion}{sf polygon of the study area (i.e. the sea!)}

\item{areas}{sf polygons among which to calculate distances. e.g. protected areas}

\item{units}{character string for output distance matrix. Either "m" or "km" (by default).}

\item{crs}{target coordinate reference system: object of class crs, or input string for st_crs}
}
\value{
distance matrix
}
\description{
in_sea_distance
}
\examples{
\dontrun{
require(MarConsNetData)
bioregion <- get_bioregion()
areas <- data_CPCAD_areas(bioregion,zones=FALSE) |>
  dplyr::mutate(area=sf::st_area(geoms))
distkm <- calc_in_sea_distance(cellsize=100000,bioregion,areas)
}
}
