% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sample_coverage.R
\name{calc_sample_coverage}
\alias{calc_sample_coverage}
\title{Calculate Chao's Sample Coverage Estimator}
\usage{
calc_sample_coverage(x, permutations = 100)
}
\arguments{
\item{x}{species by sample abundance matrix or dataframe. Rows represent samples and columns represent species.}

\item{permutations}{integer number of permutations to calculate sample coverage.}
}
\value{
A list containing the sample coverage matrix, means, and standard deviations.
}
\description{
This function calculates sample coverage for a given species by sample abundance matrix or dataframe. Uses equation 4a from Chao, A. and Jost, L. (2012), Coverage-based rarefaction and extrapolation: standardizing samples by completeness rather than size. Ecology, 93: 2533-2547. https://doi.org/10.1890/11-1952.1
}
\examples{
data(BCI,package = "vegan")
SC <- calc_sample_coverage(BCI[1:30,1:50])
ggplot(as.data.frame(SC[-1]), aes(x = N, y = means)) +
  geom_line() +
  geom_ribbon(aes(ymin = means - sd, ymax = means + sd), alpha = 0.2) +
  labs(title = "Sample Coverage", x = "Sample Size", y = "Coverage") +
  theme_minimal()

}
