% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_species_names.R
\name{clean_species_names}
\alias{clean_species_names}
\title{Clean scientific species names}
\usage{
clean_species_names(scientific_names)
}
\arguments{
\item{scientific_names}{A character vector of scientific species names.}
}
\value{
A character vector containing the cleaned scientific species names.
}
\description{
Standardizes scientific species names by replacing underscores with spaces
and formatting the first part of the name with title case.
}
\examples{
clean_species_names(c("asterias_rubens", "cancer borealis"))

}
