% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_analysis.R
\name{climate_analysis}
\alias{climate_analysis}
\title{Climate Change Analysis}
\usage{
climate_analysis(trend = NULL, itp = indicator_to_plot)
}
\arguments{
\item{trend}{trend column in indicator_to_plot}

\item{itp}{indicator_to_plot data frame, likely from the targets file}
}
\value{
a string interpreting results of indicators from a climate change lens
}
\description{
This function creates a basic interpretation of the impact of climate change
on indicators. If:
\enumerate{
\item There is a statistically significant change in results between inside
and outside comparison, we determine the change between the two could be a
result of protection efforts in the protected area
\item There is not a statistically significant change in results between inside
and outside comparison and :
A) The trend is going in the opposite direction that we would expect and
statistically significant,we point out that according to our analysis
climate change could be having the opposite impact than we could expect
B) The trend is not statistically significant, we  explain that climate
change does not seem to be having a significant impact on that particular
indicator and
C) The trend is going in the direction that we would expect and
statistically significant,we point out that according to our analysis
climate change could be having a significant impact on that indicator
}
}
