% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_report.R
\name{create_data_report}
\alias{create_data_report}
\title{Creates reports of Project Planning Tool data}
\usage{
create_data_report(datas = NULL, combine = FALSE, destdir = ".")
}
\arguments{
\item{datas}{a data frame or list of data frames obtained from \code{get_project_data()}}

\item{combine}{a Boolean indicating if the \code{datas} should be combined
into a single report. If \code{combine=FALSE} each item in the \code{datas}
argument will be an individual reports}

\item{destdir}{parameter indicating where to save the html report}
}
\description{
This function creates reports of Project Planning Tool
data
}
\examples{
\dontrun{
library(MarConsNetData)
library(MarConsNetAnalysis)
d <- get_project_data(ids=c(1093, 642))
create_data_report(datas=d)
}
}
