% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_coverage.R
\name{ind_coverage}
\alias{ind_coverage}
\title{Spatial coverage of x in y}
\usage{
ind_coverage(x, y, intersection = FALSE, proportion = FALSE)
}
\arguments{
\item{x}{sf polygon whose area acts as the numerator in the fraction (e.g. protected areas)}

\item{y}{sf polygon whose area acts as the denominator in the fraction (e.g. bioregion)}

\item{intersection}{logical that determines if the area of the spatial intersection of \code{x} and \code{y} should be used as the numerator (TRUE) or, by default, if the area of \code{x} should be used as the numerator.}

\item{proportion}{Logical that determines if this function returns a proportion (TRUE) or, by default (FALSE), a percentage (i.e. proportion X 100)}
}
\value{
numeric
}
\description{
Calculates the percent or proportion of y (e.g. bioregion) that is spatially covered by x (e.g. protected areas)
}
\examples{
require(MarConsNetData)
bioregion <- data_bioregion()
areas <- data_bioregion()
ind_coverage(x=areas, y=bioregion)

}
