% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_ProtConn.R
\name{ind_ProtConn}
\alias{ind_ProtConn}
\title{ProtConn}
\usage{
ind_ProtConn(
  distkm,
  dkm,
  bioregion,
  area = NULL,
  stepcutoff = 3,
  probcutoff = 1e-04,
  returns = "PC"
)
}
\arguments{
\item{distkm}{distance matrix in km. Can be created by the 'in_sea_distance()' in this package}

\item{dkm}{median dispersal distance in km for the negative exponential dispersal kernal (i.e. \code{p <- 1-pexp(distkm,log(2)/dkm)})}

\item{bioregion}{sf polygon of the study area (i.e. the sea!)}

\item{area}{output of data_CPCAS_areas}

\item{stepcutoff}{Maximum length of the paths (i.e. stepping stones) that are considered. Default is 3 (i.e. start -> stepping stone -> end) for a single stepping stone. If negative, no cutoff is used but will greatly increase computational requirements.}

\item{probcutoff}{Minimum probability of a connection between two areas. Default is 0.0001, i.e. any probability lower than 1 in 10000 is considered unconnected.}

\item{returns}{string describing what the function should return. By default, "PC", the function returns the ProtConn value. To return a data frame with the edge list of the network use "EL".}
}
\value{
numeric
}
\description{
ProtConn
}
\examples{
\dontrun{
require(MarConsNetData)
bioregion <- data_bioregion()
areas <- data_CPCAD_areas(bioregion,zones=FALSE) |>
  dplyr::mutate(area=sf::st_area(geoms))
distkm <- calc_in_sea_distance(cellsize=100000,bioregion,areas)
PC <- ind_ProtConn(distkm,dkm=100, area=areas)
}
}
