% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_rv_abundance.R
\name{ind_rv_abundance}
\alias{ind_rv_abundance}
\title{Calculate species abundance from the RV survey}
\usage{
ind_rv_abundance(
  species = "HADDOCK",
  area = "Western/Emerald Banks Conservation Area (Restricted Fisheries Zone)",
  MPAs = NULL,
  data = NULL
)
}
\arguments{
\item{species}{character indicating which species to analyze in the RV}

\item{area}{area of interest for species}

\item{MPAs}{an object of class "sf", likely from data_CPCAD_areas that contains
the name and geoms of the protected areas}

\item{data}{GSSPECIES from get_data in the Mar.datawrangling package}
}
\value{
\dontrun{
rva <- ind_rv_abundance()
}
}
\description{
This function calculates the average species catch per standardized tow
to 1.75 km for each year.
}
