% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indicator_weights}
\alias{indicator_weights}
\title{Indicator Weights and Categorizations}
\format{
A data frame with 129 rows and 5 variables:
\describe{
\item{\code{indicator_name}}{character the name of the indicator}
\item{\code{indicator_bin}}{character the name of the indicator bin.}
\item{\code{applicability}}{character the geographic applicability of the indicator}
\item{\code{objective}}{character the name of the EBM objective of the bin}
\item{\code{weight}}{double the relative weight}
}
}
\usage{
indicator_weights
}
\description{
The relative weights and categorizations (i.e. EBM Pillars and Objectives, and indicator bins) for all indicators.Indicators that are part of more than one bin should be listed multiple times.
}
\details{
DETAILS
}
\keyword{datasets}
