% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_data.R
\name{map_data}
\alias{map_data}
\title{Creates map of Project Planning Tool data}
\usage{
map_data(datas = NULL)
}
\arguments{
\item{datas}{a data frame or list of data frames obtained from \code{get_project_data()}}
}
\description{
This function creates a plot of the location of
Project Planning Tool data
}
\examples{
\dontrun{
library(MarConsNetData)
library(MarConsNetAnalysis)
d <- get_project_data(ids=c(1093, 642))
map_app_data(datas=d)
}
}
