% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_azmp_physical.R
\name{plot_azmp_physical}
\alias{plot_azmp_physical}
\title{Plot Physical AZMP}
\usage{
plot_azmp_physical(
  mpa = MPAs,
  area = "Western/Emerald Banks Conservation Area (Restricted Fisheries Zone)",
  type = "surface",
  dataframe = FALSE,
  parameter = "temperature"
)
}
\arguments{
\item{mpa}{an object of class "sf", likely from data_CPCAD_areas that contains
the name and geoms of the protected areas}

\item{area}{area of interest (from the NAME_E column of MPAs)}

\item{type}{surface or bottom}

\item{dataframe}{FALSE a Boolean indicating if a data frame is returned or not}

\item{parameter}{a character indicating which parameter to measure}
}
\description{
Plot Physical AZMP
}
