% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fishing_abundance.R
\name{plot_fishing_abundance}
\alias{plot_fishing_abundance}
\title{Plot leaflet raster of fishing effort}
\usage{
plot_fishing_abundance(area = "WEBCA", CPCAD = NULL, indicator = FALSE)
}
\arguments{
\item{area}{a name of an area of which to obtain the objectives from. Options include
\code{st_Anns_Bank_MPA}, \code{musquash_MPA}, \code{gully_MPA}, \code{WEBCA}}

\item{CPCAD}{a "sf" "dataframe" object from \code{\link[=data_CPCAD_areas]{data_CPCAD_areas()}}}

\item{indicator}{a Boolean indicating}
}
\value{
a "sf" "dataframe" object
}
\description{
FIXME
}
\examples{
\dontrun{
bioregion <- data_bioregion()
CPCAD <- data_CPCAD_areas(bioregion,  zones = FALSE)
plot_fishing_abundance <- function(area="WEBCA", CPCAD=CPCAD)
}

}
