% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map}
\alias{plot_map}
\title{Creates map of Project Planning Tool data}
\usage{
plot_map(datas = NULL)
}
\arguments{
\item{datas}{a data frame or list of data frames obtained from \code{get_project_data()}}
}
\description{
This function creates a plot of the location of
Project Planning Tool data
}
\examples{
\dontrun{
library(MarConsNetData)
library(MarConsNetAnalysis)
d <- get_project_data(ids=c(1093, 642))
plot_map(datas=d)
}
}
