% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rv_abundance.R
\name{plot_rv_abundance}
\alias{plot_rv_abundance}
\title{Plot Average number of species per tow per year}
\usage{
plot_rv_abundance(abundance = NULL)
}
\arguments{
\item{abundance}{data.frame likely from ind_rv_abundance}
}
\description{
This function plots the average species catch per standardized tow
to 1.75 km for each year.
}
\examples{
\dontrun{
rva <- ind_rv_abundance()
plot_rv_abundance(rva)
}
}
