% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trend_status.R
\name{plot_trend_status}
\alias{plot_trend_status}
\title{Plot Physical AZMP}
\usage{
plot_trend_status(
  mpa = NULL,
  area = "Western/Emerald Banks Conservation Area (Restricted Fisheries Zone)",
  type = "surface",
  dataframe = FALSE,
  parameter = "temperature",
  outside = FALSE
)
}
\arguments{
\item{mpa}{an object of class "sf", likely from data_CPCAD_areas that contains
the name and geoms of the protected areas}

\item{area}{area of interest (from the NAME_E column of MPAs)}

\item{type}{surface or bottom}

\item{dataframe}{FALSE a Boolean indicating if a data frame is returned or not}

\item{parameter}{a character indicating which parameter to measure}

\item{outside}{a Boolean indicating if an outside comparison is happening}
}
\description{
Plot Physical AZMP
}
