% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProtConn.R
\name{ProtConn}
\alias{ProtConn}
\title{ProtConn}
\usage{
ProtConn(distkm, dkm)
}
\arguments{
\item{distkm}{distance matrix in km. Can be created by the 'in_sea_distance()' in this package}

\item{dkm}{median dispersal distance in km for the negative exponential dispersal kernal (i.e. \code{p <- 1-pexp(distkm,log(2)/dkm)})}
}
\value{
numeric
}
\description{
ProtConn
}
\examples{
require(MarConsNetData)
bioregion <- get_bioregion()
areas <- get_CPCAD_areas(bioregion,zones=FALSE) |>
  dplyr::mutate(area=sf::st_area(geoms))
distkm <- in_sea_distance(cellsize=100000,bioregion,areas)
PC <- ProtConn(distkm,dkm=100)
}
