% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_indicator.R
\name{score_indicator}
\alias{score_indicator}
\title{Title}
\usage{
score_indicator(
  data,
  indicator_var_name = NA,
  indicator,
  type = NA,
  units = NA,
  scoring = NA,
  project = NA,
  climate = FALSE,
  design_target = FALSE,
  crs = 4326,
  latitude = "latitude",
  longitude = "longitude",
  year = "year",
  other_nest_variables = NA
)
}
\arguments{
\item{other_nest_variables}{}
}
\description{
Title
}
