% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analysis}
\alias{analysis}
\title{Analyse Status and Trends of indicators}
\usage{
analysis(bi = binned_indicators)
}
\arguments{
\item{bi}{binned_indicators data frame likely found in the
data folder of the MarConsNetAnalysis package}
}
\value{
data frame with trends and status'
}
\description{
This function automatically calculates status and trends
for a list of indicators used in the Marine Conservation
Target (MCT) app
}
