% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_letter_grade.R
\name{calc_letter_grade}
\alias{calc_letter_grade}
\title{Calculate letter grades}
\usage{
calc_letter_grade(scores)
}
\arguments{
\item{scores}{vector of numeric scores to be converted to letter grades}
}
\value{
vector of letter grades
}
\description{
Calculate letter grades
}
\examples{

scores <- c(10, 20, 30, 40, 50, 60, 70, 80, 90, 100)

calc_letter_grade(scores)

}
