% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_placeholder.R
\name{ind_placeholder}
\alias{ind_placeholder}
\title{Create example FAKE data}
\usage{
ind_placeholder(
  ind_name = paste0("placeholder_", round(runif(1, 1, 100))),
  areas,
  areaName = "NAME_E"
)
}
\arguments{
\item{areaName}{}
}
\description{
Create example FAKE data
}
