% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_ProtConn.R
\name{ind_ProtConn}
\alias{ind_ProtConn}
\title{ProtConn}
\usage{
ind_ProtConn(distkm, dkm, bioregion)
}
\arguments{
\item{distkm}{distance matrix in km. Can be created by the 'in_sea_distance()' in this package}

\item{dkm}{median dispersal distance in km for the negative exponential dispersal kernal (i.e. \code{p <- 1-pexp(distkm,log(2)/dkm)})}

\item{bioregion}{sf polygon of the study area (i.e. the sea!)}
}
\value{
numeric
}
\description{
ProtConn
}
\examples{
require(MarConsNetData)
bioregion <- data_bioregion()
areas <- data_CPCAD_areas(bioregion,zones=FALSE) |>
  dplyr::mutate(area=sf::st_area(geoms))
distkm <- calc_in_sea_distance(cellsize=100000,bioregion,areas)
PC <- ind_ProtConn(distkm,dkm=100)
}
