% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{_PACKAGE}
\name{MarConsNetAnalysis-package}
\alias{MarConsNetAnalysis-package}
\title{A Package for Analyzing MPA Data}
\description{
The \code{MarConsNetAnalysis} package is intended to analyze the
real data  that was located and transformed using the
\code{MarConsNetData} package.
}
