% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_flowerplot.R
\name{plot_flowerplot}
\alias{plot_flowerplot}
\title{Plot a flowerplot}
\usage{
plot_flowerplot(
  df,
  grouping = "grouping",
  labels = "labels",
  score = "score",
  weight = "weight"
)
}
\arguments{
\item{weight}{}
}
\description{
Plot a flowerplot
}
