% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_flowerplot.R
\name{plot_flowerplot}
\alias{plot_flowerplot}
\title{Plot a flowerplot}
\usage{
plot_flowerplot(
  df,
  grouping = "grouping",
  labels = "labels",
  score = "score",
  weight = "weight",
  title = unique(df[["area_name"]]),
  max_score = 100,
  min_score = 0,
  bintextsize = 3,
  zeroline = FALSE
)
}
\arguments{
\item{df}{data.frame with indicator scores, grouping, labels, and relative weight}

\item{grouping}{character string for the name of the grouping column in \code{df}}

\item{labels}{character string for the name of the labels column in \code{df}}

\item{score}{character string for the name of the score column in \code{df}}

\item{weight}{character string for the name of the weight column in \code{df}}

\item{title}{Defaults to the unique value of the \code{area_name} column of the \code{df}, but can take any character value. Alternatively, use \code{FALSE} to avoid having a title.}

\item{max_score}{numeric value for the maximum possible value of the scale (i.e. petal length). Default is 100}

\item{min_score}{numeric value for the minimum possible value of the scale (i.e. petal length). Default is 0}

\item{bintextsize}{numeric value for the size of the text in the bins. Default is 3}

\item{zeroline}{logical value to add a zero line to the plot. Default is \code{FALSE}}
}
\value{
plot
}
\description{
Plot a flowerplot
}
\examples{

indicatorbins <- data.frame(grouping=rep(c("Biodiversity",
                                           "Habitat",
                                           "Productivity"),
                                         times=c(3,5,3)),
                            labels=c("Genetic Diversity",
                                     "Species Diversity",
                                     "Functional Diversity",

                                     "Environmental Representativity",
                                     "Key Fish Habitat",
                                     "Connectivity",
                                     "Uniqueness",
                                     "Threats to Habitat",

                                     "Biomass Metrics",
                                     "Structure and Function",
                                     "Threats to Productivity"),
                            score=runif(11,55,100),
                            weight=1,
                            area_name = "Random Example MPA")

plot_flowerplot(indicatorbins)


}
