% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_indicator.R
\name{plot_indicator}
\alias{plot_indicator}
\title{Plot Indicator}
\usage{
plot_indicator(
  data,
  indicator,
  units,
  id,
  plot_type,
  year,
  indicator_var_name,
  scoring,
  areaID,
  areas,
  bin_width
)
}
\arguments{
\item{data}{data processed in \code{assess_indicator()}}

\item{indicator}{Character. Human-readable indicator name.}

\item{units}{Character. Units of measurement.}

\item{id}{Equivalent to \code{areaID} in \code{\link[=process_indicator]{process_indicator()}}}

\item{plot_type}{Character. Plot type(s) to generate
(e.g., \code{"time-series"}, \code{"map-species"}).}

\item{year}{Character. Name of year column in dataset.}

\item{indicator_var_name}{Character. Name of the column in \code{data} containing
the indicator values.}

\item{scoring}{Character. Method used to score the indicator. Options include:
\itemize{
\item \code{"desired trend"} – Uses regression to score based on increase,
decrease, or stability.
\item \code{"representation"} – Scores based on number of features that
intersect with a protected area.
\item \code{"coverage"} – Scores based on \% overlap of a feature within
protected areas.
\item \code{"median"} – Uses median value of raster or continuous data, rescaled
to 0–100.
\item \code{"control site comparison"} – Compares trends inside vs. outside
managed areas.
}}

\item{areaID}{Character. Name of the area identifier column in \code{areas}.}

\item{areas}{A spatial or tabular object defining assessment areas.}

\item{bin_width}{Numeric. Bin width used for plotting or scoring.}
}
\value{
a plot
}
\description{
Plot Indicator
}
