% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_indicator.R
\name{plot_indicator}
\alias{plot_indicator}
\title{Plot Indicator}
\usage{
plot_indicator(
  data,
  indicator,
  units,
  id,
  plot_type,
  year,
  indicator_var_name,
  scoring,
  areaID,
  areas
)
}
\arguments{
\item{data}{data processed in \code{assess_indicator()}}

\item{indicator}{Character. Indicator name or short description.}

\item{units}{Character. Units of the indicator (e.g., "kg/ha", "\% cover").}

\item{id}{Equivalent to \code{areaID} in \code{\link[=process_indicator]{process_indicator()}}}

\item{scoring}{Character. Method used to score the indicator. Options include:
\itemize{
\item \code{"desired trend"} – Uses regression to score based on increase,
decrease, or stability.
\item \code{"representation"} – Scores based on \% overlap of feature within
protected areas.
\item \code{"median"} – Uses median value of raster or continuous data, rescaled
to 0–100.
\item \code{"control site comparison"} – Compares trends inside vs. outside
managed areas.
}}

\item{areaID}{Character. The name of the column in \code{areas} and \code{data} used to
match records to geographic areas.}

\item{areas}{An \code{sf} polygon object defining the areas of interest (e.g.,
conservation areas, regions). Used for spatial joins and reporting.}
}
\value{
a plot
}
\description{
Plot Indicator
}
