% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_indicator.R
\name{process_indicator}
\alias{process_indicator}
\title{Title}
\usage{
process_indicator(
  data,
  indicator_var_name = NA,
  indicator,
  type = NA,
  units = NA,
  scoring = NA,
  project = NA,
  climate = FALSE,
  design_target = FALSE,
  crs = 4326,
  latitude = "latitude",
  longitude = "longitude",
  year = "year",
  other_nest_variables = NA,
  areas = NA,
  areaID = "NAME_E"
)
}
\arguments{
\item{areaID}{}
}
\description{
Title
}
