% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_paths.R
\name{tar_paths}
\alias{tar_paths}
\title{Convenience that helps generate relevant paths to run the 'targets' pipeline}
\usage{
tar_paths(path = "onedrive")
}
\arguments{
\item{path}{character string for the path to return. Defaults to "onedrive"
which will return the shared onedrive target store (i.e. the data). Both "pkg"
or "inst" will return the path to the "_targets.R" file (i.e. the instructions)
in the installed package directory or the inst directory in your local git repo
respectively.}
}
\value{
character string
}
\description{
Convenience that helps generate relevant paths to run the 'targets' pipeline
}
\examples{

onedrive_store <- tar_paths()

}
